move css files into style/ folder
