+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { HashRouter as Router } from 'react-router-dom'; 
-import './App.css';
+import './style/App.css';
 import Content from './Content.js';
 
 class App extends Component {
